requirements.txt
Flask==2.1.1
Werkzeug==2.0.2
requests==2.26.0
beautifulsoup4==4.10.0
python-dotenv==0.19.1
Gunicorn==20.1.0
